requirements.txt
Django>=4.0
djangorestframework>=3.12
djangorestframework-simplejwt>=4.8
psycopg2-binary>=2.9  
django-redis>=5.0  
